+++ .cypress/cypress/integration/peterborough.js
@@ -13,22 +13,24 @@ describe('new report form', function() {
 
   it('is hidden when emergency option is yes', function() {
     cy.get('select:eq(4)').select('Fallen branch');
+    cy.get('.js-reporting-page--next:visible').click();
     cy.get('#form_emergency').select('yes');
-    cy.get('#js-category-stopper').should('contain', 'Please phone customer services to report this problem.');
-    cy.get('.js-hide-if-invalid-category').should('be.hidden');
+    cy.get('.js-post-category-messages:visible').should('contain', 'Please phone customer services to report this problem.');
+    cy.get('.js-reporting-page--next:visible').should('be.disabled');
     cy.get('#form_emergency').select('no');
-    cy.get('#js-category-stopper').should('not.contain', 'Please phone customer services to report this problem.');
-    cy.get('.js-hide-if-invalid-category').should('be.visible');
+    cy.get('.js-post-category-messages:visible').should('not.contain', 'Please phone customer services to report this problem.');
+    cy.get('.js-reporting-page--next:visible').should('not.be.disabled');
   });
 
   it('is hidden when private land option is yes', function() {
     cy.get('select:eq(4)').select('Fallen branch');
+    cy.get('.js-reporting-page--next:visible').click();
     cy.get('#form_private_land').select('yes');
-    cy.get('#js-category-stopper').should('contain', 'The council do not have powers to address issues on private land.');
-    cy.get('.js-hide-if-invalid-category').should('be.hidden');
+    cy.get('.js-post-category-messages:visible').should('contain', 'The council do not have powers to address issues on private land.');
+    cy.get('.js-reporting-page--next:visible').should('be.disabled');
     cy.get('#form_private_land').select('no');
-    cy.get('#js-category-stopper').should('not.contain', 'The council do not have powers to address issues on private land.');
-    cy.get('.js-hide-if-invalid-category').should('be.visible');
+    cy.get('.js-post-category-messages:visible').should('not.contain', 'The council do not have powers to address issues on private land.');
+    cy.get('.js-reporting-page--next:visible').should('not.be.disabled');
   });
 
   it('correctly changes the asset select message', function() {